test: deploy xudt contract binary and rce cell in tests
