+++ test/src/spec/mod.rs
@@ -1,14 +1,14 @@
 mod dao;
 mod plugin;
 mod rpc;
-mod sudt;
+mod udt;
 mod util;
 mod wallet;
 
 pub use dao::*;
 pub use plugin::*;
 pub use rpc::*;
-pub use sudt::*;
+pub use udt::*;
 pub use util::*;
 pub use wallet::*;
